rename prompts to prompt
