+++ src/dbally/iql_generator/iql_prompt_template.py
@@ -1,9 +1,9 @@
 from typing import List
 
 from dbally.exceptions import DbAllyError
-from dbally.prompts import PromptTemplate
-from dbally.prompts.elements import FewShotExample
-from dbally.prompts.template import PromptFormat
+from dbally.prompt import PromptTemplate
+from dbally.prompt.elements import FewShotExample
+from dbally.prompt.template import PromptFormat
 from dbally.views.exposed_functions import ExposedFunction
