+++ src/dbally/view_selection/view_selector_prompt_template.py
@@ -1,8 +1,8 @@
 from typing import Dict, List
 
-from dbally.prompts import PromptTemplate
-from dbally.prompts.elements import FewShotExample
-from dbally.prompts.template import PromptFormat
+from dbally.prompt import PromptTemplate
+from dbally.prompt.elements import FewShotExample
+from dbally.prompt.template import PromptFormat
 
 
 class ViewSelectionPromptFormat(PromptFormat):
